210728-2

+++ src/Youma.py
@@ -2,10 +2,11 @@
 import os
 import json
 from traceback import format_exc
+
 from Class.MyHandler import Handler
 from Class.MyEnum import ScrapeStatusEnum
 from Class.MyLogger import Logger
-from Class.MyJav import JavFile, JavModel
+from Class.MyJav import JavModel
 from Class.MyError import TooManyDirectoryLevelsError, SpecifiedUrlError
 from Functions.Progress.User import choose_directory
 from Functions.Metadata.Genre import better_dict_youma_genres
@@ -15,9 +16,12 @@ from Functions.Web.Javlibrary import scrape_from_library
 
 #  main开始
 from Javdb import scrape_from_db
+from JsonUtility import read_json_to_dict
 
-print('1、避开高峰期，整理速度可能很慢。\n'
-      '2、若一直打不开网站，请在ini中更新对应网址\n')
+print('1、这是测试版，仅对【有码】影片整理有效！！！有报错请联系！！！'
+      '2、请自行将【无码】【素人】的影片移至他处，虽然也能整理部分，但可能出错！！！'
+      '   例如报错”新的特征需添加至【特征对照表】“，因为暂时只收纳了有码特征！！！'
+      '3、”繁简“功能暂不支持，默认简体')
 
 # region（1）读取配置文件
 print('正在读取ini中的设置...', end='')
@@ -34,7 +38,7 @@ print('\n读取ini文件成功!\n')
 # 路径分隔符: 当前系统的路径分隔符 windows是“\”，linux和mac是“/”
 sep = os.sep
 # 当前程序文件夹 所处的 父文件夹路径
-dir_project_father = os.path.dirname(os.getcwd())
+dir_pwd_father = os.path.dirname(os.getcwd())
 # arzon通行证: 如果需要从arzon获取日语简介，需要先获得合法的arzon网站的cookie，用于通过成人验证。
 cookie_arzon = handler.get_last_arzon_cookie()
 # 优化特征的字典
@@ -67,9 +71,9 @@ while not input_key:
         # region （3.2.1）当前文件夹内包含jav及字幕文件的状况: 有多少视频，其中多少jav，同一车牌多少cd，当前文件夹是不是独立文件夹
         # （3.2.1.1）什么文件都没有 | 当前目录是之前已归类的目录，无需处理 | 判断这一层文件夹中有没有nfo
         if not list_sub_files \
-                or '归类完成' in dir_current[len(dir_choose):] \
-                or handler.judge_skip_exist_nfo(list_sub_files):
-            continue    # dir_current[len(dir_choose):] 当前所处文件夹 相对于 所选文件夹 的路径，主要用于报错
+                or '归类完成' in dir_current[len(dir_choose):]:
+            # or handler.judge_skip_exist_nfo(list_sub_files):
+            continue  # dir_current[len(dir_choose):] 当前所处文件夹 相对于 所选文件夹 的路径，主要用于报错
 
         # （3.2.1.2）判断文件是不是字幕文件，放入dict_subtitle_file中，字幕文件和车牌对应关系 {'c:\a\abc_123.srt': 'abc-123'}
         handler.init_dict_subtitle_file(list_sub_files)
@@ -79,10 +83,10 @@ while not input_key:
         if not list_jav_files:
             continue
         # （3.2.1.5）判定当前所处文件夹是否是独立文件夹，独立文件夹是指该文件夹仅用来存放该影片，而不是大杂烩文件夹，是后期移动剪切操作的重要依据
-        # 错误：JavFile.Is_in_separate_folder = handler.judge_separate_folder(len(list_jav_files), list_sub_dirs)
-        # 我在主程序里修改JavFile.Is_in_separate_folder并不会成功，你知道为什么吗？
+        # 错误：JavFile.Bool_in_separate_folder = handler.judge_separate_folder(len(list_jav_files), list_sub_dirs)
+        # 我在主程序里修改JavFile.Bool_in_separate_folder并不会成功，你知道为什么吗？
         handler.judge_separate_folder(len(list_jav_files), list_sub_dirs)
-        # is_in_separate_folder是类属性，不是实例属性，修改类属性会将list_jav_files中的所有jav_file的is_in_separate_folder同步
+        # Bool_in_separate_folder是类属性，不是实例属性，修改类属性会将list_jav_files中的所有jav_file的Bool_in_separate_folder同步
         # （3.2.1.6）处理“集”的问题，（1）所选文件夹总共有多少个视频文件，包括非jav文件，主要用于显示进度（2）同一车牌有多少cd，用于cd2...命名
         handler.count_num_and_no(list_jav_files)
         # endregion
@@ -94,65 +98,74 @@ while not input_key:
                 # 当前进度
                 print(f'>> [{jav_file.No}/{handler.sum_videos_in_choose_dir}]:{jav_file.Name}')
                 print(f'    >发现车牌: {jav_file.Car}')
-                jav_model = JavModel(jav_file.Car)
                 logger.path_relative = jav_file.Path[len(dir_choose):]  # 影片的相对于所选文件夹的路径，用于报错
                 # endregion
 
-                # region（3.2.2.2）从javdb获取信息
-                status, genres_db = scrape_from_db(jav_file, jav_model, handler.url_db, handler.proxy_db)
-                temp = status == ScrapeStatusEnum.db_not_found
-                if status == ScrapeStatusEnum.db_not_found:
-                    logger.record_warn(f'javdb找不到该车牌的信息: {jav_file.Car}，')
-                # 优化genres_db
-                genres_db = [dict_db_genres[i] for i in genres_db if dict_db_genres[i] != '删除']
-                # endregion
+                dir_prefs_jsons = f'{dir_pwd_father}{sep}【重要须备份】已整理的jsons{sep}{jav_file.Pref}{sep}'
+                path_json = f'{dir_prefs_jsons}{jav_file.Car}.json'
+                if os.path.exists(path_json):
+                    jav_model = JavModel(**read_json_to_dict(path_json))
+                    genres = jav_model.Genres
+                    print(f'    >从本地json读取元数据: {path_json}')
+                else:
+                    jav_model = JavModel()
+                    # region（3.2.2.2）从javdb获取信息
+                    status, genres_db = scrape_from_db(jav_file, jav_model, handler.url_db, handler.proxy_db)
+                    temp = status == ScrapeStatusEnum.db_not_found
+                    if status == ScrapeStatusEnum.db_not_found:
+                        logger.record_warn(f'javdb找不到该车牌的信息: {jav_file.Car}，')
+                    # 优化genres_db
+                    genres_db = [dict_db_genres[i] for i in genres_db if dict_db_genres[i] != '删除']
+                    # endregion
 
-                # region（3.2.2.3）从javlibrary获取信息
-                status, genres_library = scrape_from_library(jav_file, jav_model, handler.url_library,
-                                                             handler.proxy_library)
-                if status == ScrapeStatusEnum.library_not_found:
-                    logger.record_warn(f'javlibrary找不到该车牌的信息: {jav_file.Car}，')
-                elif status == ScrapeStatusEnum.library_multiple_search_results:
-                    logger.record_warn(f'javlibrary搜索到同车牌的不同视频: {jav_file.Car}，')
-                # 优化genres_library
-                genres_library = [dict_library_genres[i] for i in genres_library if dict_library_genres[i] != '删除']
-                # endregion
+                    # region（3.2.2.3）从javlibrary获取信息
+                    status, genres_library = scrape_from_library(jav_file, jav_model, handler.url_library,
+                                                                 handler.proxy_library)
+                    if status == ScrapeStatusEnum.library_not_found:
+                        logger.record_warn(f'javlibrary找不到该车牌的信息: {jav_file.Car}，')
+                    elif status == ScrapeStatusEnum.library_multiple_search_results:
+                        logger.record_warn(f'javlibrary搜索到同车牌的不同视频: {jav_file.Car}，')
+                    # 优化genres_library
+                    genres_library = [dict_library_genres[i] for i in genres_library if dict_library_genres[i] != '删除']
+                    # endregion
 
-                if not jav_model.Javdb and not jav_model.Javlibrary:
-                    logger.record_fail(f'Javdb和Javlibrary都找不到该车牌信息: {jav_file.Car}，')
-                    continue  # 结束对该jav的整理
+                    if not jav_model.Javdb and not jav_model.Javlibrary:
+                        logger.record_fail(f'Javdb和Javlibrary都找不到该车牌信息: {jav_file.Car}，')
+                        continue  # 结束对该jav的整理
 
-                # region（3.2.2.4）前往javbus查找【封面】【系列】【特征】
-                status, genres_bus = scrape_from_bus(jav_file, jav_model, handler.url_bus, handler.proxy_bus)
-                if status == ScrapeStatusEnum.bus_multiple_search_results:
-                    logger.record_warn(f'部分信息可能错误，javbus搜索到同车牌的不同视频: {jav_file.Car_id}，')
-                elif status == ScrapeStatusEnum.bus_not_found:
-                    logger.record_warn(f'javbus有码找不到该车牌的信息: {jav_file.Car_id}，')
-                # 优化genres_bus
-                genres_bus = [dict_bus_genres[i] for i in genres_bus
-                              if not i.startswith('AV OP')
-                              or not i.startswith('AVOP')
-                              or dict_bus_genres[i] != '删除']
-                # endregion
+                    # region（3.2.2.4）前往javbus查找【封面】【系列】【特征】
+                    status, genres_bus = scrape_from_bus(jav_file, jav_model, handler.url_bus, handler.proxy_bus)
+                    if status == ScrapeStatusEnum.bus_multiple_search_results:
+                        logger.record_warn(f'部分信息可能错误，javbus搜索到同车牌的不同视频: {jav_file.Car_id}，')
+                    elif status == ScrapeStatusEnum.bus_not_found:
+                        logger.record_warn(f'javbus有码找不到该车牌的信息: {jav_file.Car_id}，')
+                    # 优化genres_bus
+                    genres_bus = [dict_bus_genres[i] for i in genres_bus
+                                  if not i.startswith('AV OP')
+                                  or not i.startswith('AVOP')
+                                  or dict_bus_genres[i] != '删除']
+                    # endregion
 
-                # region（3.2.2.5）arzon找简介
-                status, cookie_arzon = scrape_from_arzon(jav_file, jav_model, cookie_arzon, handler.proxy_arzon)
-                url_search_arzon = f'https://www.arzon.jp/itemlist.html?t=&m=all&s=&q={jav_file.Car_id.replace("-", "")}'
-                if status == ScrapeStatusEnum.arzon_exist_but_no_plot:
-                    logger.record_warn(f'找不到简介，尽管arzon上有搜索结果: {url_search_arzon}，')
-                elif status == ScrapeStatusEnum.arzon_not_found:
-                    logger.record_warn(f'找不到简介，影片被arzon下架: {url_search_arzon}，')
-                elif status == ScrapeStatusEnum.interrupted:
-                    logger.record_warn(f'访问arzon失败，需要重新整理该简介: {url_search_arzon}，')
-                # endregion
+                    # region（3.2.2.5）arzon找简介
+                    status, cookie_arzon = scrape_from_arzon(jav_file, jav_model, cookie_arzon, handler.proxy_arzon)
+                    url_search_arzon = f'https://www.arzon.jp/itemlist.html?t=&m=all&s=&q={jav_file.Car_id.replace("-", "")}'
+                    if status == ScrapeStatusEnum.arzon_exist_but_no_plot:
+                        logger.record_warn(f'找不到简介，尽管arzon上有搜索结果: {url_search_arzon}，')
+                    elif status == ScrapeStatusEnum.arzon_not_found:
+                        logger.record_warn(f'找不到简介，影片被arzon下架: {url_search_arzon}，')
+                    elif status == ScrapeStatusEnum.interrupted:
+                        logger.record_warn(f'访问arzon失败，需要重新整理该简介: {url_search_arzon}，')
+                    # endregion
+                    # region（3.2.2.6）后续完善操作
+                    # 整合genres
+                    jav_model.Genres = genres = list(set(genres_db + genres_library + genres_bus))
 
-                # region（3.2.2.6）后续完善操作
-                # 整合完善genres
-                jav_model.Genres = genres = list(set(genres_db + genres_library + genres_bus))
-                if jav_file.Is_subtitle:  # 有“中字“，加上特征”中文字幕”
+                # 完善genres
+                if jav_file.Bool_subtitle:  # 有“中字“，加上特征”中文字幕”
                     genres.append('中文字幕')
-                if jav_file.Is_divulge:  # 是流出无码片，加上特征'无码流出'
+                if jav_file.Bool_divulge:  # 是流出无码片，加上特征'无码流出'
                     genres.append('无码流出')
+
                 # 完善handler.dict_for_standard
                 handler.prefect_jav_model_and_dict_for_standard(jav_file, jav_model)
                 # 完善jav_model.CompletionStatus
@@ -160,13 +173,12 @@ while not input_key:
                 # endregion
 
                 ################################################################################
-                dir_prefs_jsons = f'{dir_project_father}{sep}已整理的jsons{sep}{jav_file.Pref}{sep}'
                 if not os.path.exists(dir_prefs_jsons):
                     os.makedirs(dir_prefs_jsons)
-                path_json = f'{dir_prefs_jsons}{jav_model.Release}{jav_model.Car}.json'
-                with open(path_json, 'w', encoding='utf-8') as f:
-                    json.dump(jav_model.__dict__, f, indent=4)
-                print(f'    >保存本地json成功: {path_json}')
+                if not os.path.exists(path_json):
+                    with open(path_json, 'w', encoding='utf-8') as f:
+                        json.dump(jav_model.__dict__, f, indent=4)
+                    print(f'    >保存本地json成功: {path_json}')
 
                 # 1重命名视频
                 path_new = handler.rename_mp4(jav_file)
